requirements.txt
streamlit
openai>=1.40.0
pydantic>=2.6.0
python-dotenv
faiss-cpu
sentence-transformers
pymupdf
beautifulsoup4
requests
